+++ apps/cms/models/tag.py
@@ -1,0 +1,6 @@
+from .base import SlugBase
+
+
+class Tag(SlugBase):
+    pass
+
